Remove accidentally added unused generator, and move BigFractionToLongGenerator to the correct package
