+++ src/main/java/moze_intel/projecte/emc/generators/BigFractionToLongGenerator.java
@@ -1,4 +1,4 @@
-package moze_intel.projecte.emc.arithmetics;
+package moze_intel.projecte.emc.generators;
 
 import moze_intel.projecte.emc.generators.IValueGenerator;
 import org.apache.commons.math3.fraction.BigFraction;
