+++ src/test/java/moze_intel/projecte/emc/GraphMapperTest.java
@@ -1,6 +1,6 @@
 package moze_intel.projecte.emc;
 
-import moze_intel.projecte.emc.arithmetics.BigFractionToLongGenerator;
+import moze_intel.projecte.emc.generators.BigFractionToLongGenerator;
 import moze_intel.projecte.emc.arithmetics.HiddenBigFractionArithmetic;
 import moze_intel.projecte.emc.arithmetics.IValueArithmetic;
 import moze_intel.projecte.emc.collector.LongToBigFractionCollector;
